pyproject.toml
[project]
name = "yolo-clean"
version = "0.1.0"
description = "A clean YOLO implementation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "torch>=2.0",
    "opencv-python",
    "numpy",
    "tqdm",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "pyright",
]
debug = [
    # Deps needed to load reference yolov9 checkpoints (pickled model objects)
    "ipython",
    "matplotlib",
    "pandas",
    "seaborn",
    "scipy",
    "torchvision",
    "requests",
    "psutil",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yolo"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
